# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Trading app between exchange students. For example, Erasmus people who are leaving the country where they did their exchange experiance can't carry everything they bought because it it too expensive to send it to their country, so they can send it to new incoming Erasmus who is quite probable that they will need the same things"
  version: "1.0.0"
  title: "Swagger TradErasmus"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:9000"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /publication/{publicationCity}:
    get:
      tags:
      - "publications"
      summary: "Search publication"
      description: "Search publications in a city where you are interested to buy."
      operationId: "getPublicationByCity"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "publicationCity"
        in: "path"
        description: "Publications that match with the input string"
        required: true
        type: "integer"
        format: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Publication"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Publications not found"
      security:
      - api_key: []
      x-swagger-router-controller: "Publications"
    put:
      tags:
      - "publications"
      summary: "Updates a publication"
      description: "Updates a publication which the owner is interested to"
      operationId: "updatePublicationWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "publicationId"
        in: "path"
        description: "ID of publication that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "name"
        in: "formData"
        description: "Updated price of the publication"
        required: false
        type: "string"
      - name: "status"
        in: "formData"
        description: "Updated status of the publications"
        required: false
        type: "string"
      responses:
        "405":
          description: "Invalid input"
      security:
      - publications_auth:
        - "write:publications"
        - "read:publications"
      x-swagger-router-controller: "Publications"
...
securityDefinitions:
  publications_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost:9000/oauth/dialog"
    flow: "implicit"
    scopes:
      write:publications: "Modify publications in your account"
      read:publications: "Read your publications"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
    example:
      name: "name"
      id: 6
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
    example:
      name: "name"
      id: 1
  Publication:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "publications status in the store"
        enum:
        - "available"
        - "pending"
        - "sold"
    xml:
      name: "publications"
    example:
      photoUrls:
      - "photoUrls"
      - "photoUrls"
      name: "doggie"
      id: 0
      category:
        name: "name"
        id: 6
      tags:
      - name: "name"
        id: 1
      - name: "name"
        id: 1
      status: "available"